+++ packages/core/src/model/EditConfigModel.ts
@@ -125,6 +125,8 @@ export interface IEditConfigType {
   edgeTextMode: TextMode
   // 开启网格对齐
   snapGrid: boolean
+  isPinching: boolean
+  anchorProximityValidate: boolean
 }
 
 export type IConfigKeys = keyof IEditConfigType
@@ -186,6 +188,8 @@ const allKeys = [
   'edgeTextMultiple', // 是否支持多个边文本
   'nodeTextVertical', // 节点文本是否纵向显示
   'edgeTextVertical', // 边文本是否纵向显示
+  'isPinching', //是否是双指捏合态
+  'anchorProximityValidate', // 仅在靠近锚点时触发连接校验
 ] as const
 
 /**
@@ -202,6 +206,7 @@ export class EditConfigModel {
   @observable stopMoveGraph = false
   @observable stopScrollGraph = false
   @observable snapGrid = false
+  @observable isPinching = false
   /*********************************************************
    * 文本相关配置（全局）
    ********************************************************/
@@ -220,6 +225,7 @@ export class EditConfigModel {
   @observable edgeTextDraggable = false
   @observable edgeTextMultiple = false // 是否支持多个边文本
   @observable edgeTextVertical = false // 边文本朝向是否是纵向
+  @observable anchorProximityValidate = false // 仅在靠近锚点时触发连接校验
   /*********************************************************
    * 节点相关配置
    ********************************************************/