+++ packages/vue-node-registry/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @logicflow/vue-node-registry
 
+## 1.2.0-alpha.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @logicflow/core@2.2.0-alpha.2
+
+## 1.2.0-alpha.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @logicflow/core@2.2.0
+
+## 1.2.0-alpha.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @logicflow/core@2.2.0-alpha.0
+
+## 1.1.5
+
+### Patch Changes
+
+- fix: 修复issue反馈的问题
+  - fix: 修复extension包引入vue包问题 fix #2277
+  - Updated dependencies
+    - @logicflow/core@2.1.4
+
 ## 1.1.4
 
 ### Patch Changes